chore(ui): restructure composables into dedicated screen folders
