+++ app/src/main/java/com/example/traderapp/data/model/UserProfile.kt
@@ -1,4 +1,4 @@
-package com.example.traderapp.ui.screens.components
+package com.example.traderapp.data.model
