--- app/src/main/java/com/example/traderapp/ui/screens/components/StockItem.kt
@@ -1,2 +1,0 @@
-package com.example.traderapp.ui.screens.components
-
