+++ app/src/main/java/com/example/traderapp/ui/screens/onboarding/OnBoardingTitle.kt
@@ -1,4 +1,4 @@
-package com.example.traderapp.ui.screens.components.onboarding
+package com.example.traderapp.ui.screens.onboarding
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
