+++ app/src/main/java/com/example/traderapp/ui/screens/portfolio/PortfolioCompactSection.kt
@@ -1,4 +1,4 @@
-package com.example.traderapp.ui.screens.components
+package com.example.traderapp.ui.screens.portfolio
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.*
